Change file names and add style changes
